feat: add basic markdown to produrt-page
