+++ src/app/app.ts
@@ -59,8 +59,7 @@ export default class App {
         },
       },
       {
-        // path: `${Page.PRODUCT}/${ID_SELECTOR}`,
-        path: Page.PRODUCT,
+        path: `${Page.PRODUCT}/${ID_SELECTOR}`,
         callback: (id: string) => {
           this.main.setContent([new ProductPage().getElement()]);
         },
