+++ src/features/slider/slider.scss
@@ -33,7 +33,7 @@ $imageWidth: 575px;
   &__button {
     position: relative;
     z-index: 1;
-    // margin-right: - $imageWidth * 0.15;
+    margin: -$imageWidth * 0.12;
     width: $imageWidth * 0.05;
     height: $imageWidth;
     cursor: pointer;
@@ -66,7 +66,7 @@ $imageWidth: 575px;
     }
 
     &_next {
-      margin-right: 0;
+      // margin-right: 0;
       // margin-left: - $imageWidth * 0.15;
       transform: rotate(180deg);
     }
